fix: Run CI on PHP 8.3 and fix deprecation notices (#211)
* fix: Deprecation notices

Removed deprecation notices for dynamic properties and implicit int conversions.

* chore(ci): Run CI tests on 8.3.

CI tests now run on 8.3 too, also readme now reflects support for 8.3.

+++ tests/Rasterization/Path/ArcApproximatorTest.php
@@ -143,7 +143,7 @@ class ArcApproximatorTest extends \PHPUnit\Framework\TestCase
         $this->assertEqualsWithDelta(20, $result[$n - 1][0], 0.1);
         $this->assertEqualsWithDelta(10, $result[$n - 1][1], 0.1);
         // test some point roughly in the middle
-        $this->assertEqualsWithDelta(15, $result[$n / 2][0], 1);
-        $this->assertEqualsWithDelta(15, $result[$n / 2][1], 1);
+        $this->assertEqualsWithDelta(15, $result[(int)($n / 2)][0], 1);
+        $this->assertEqualsWithDelta(15, $result[(int)($n / 2)][1], 1);
     }
 }
